chore: moved imprint_screen to info folder
